Takudo (#38)
* update header

* part of the checkout

* show, process customer order - admin

* show user order

* add filter admin product

+++ database/migrations/2022_04_27_170457_create_products_table.php
@@ -20,7 +20,7 @@ class CreateProductsTable extends Migration
             $table->float('amount');
             $table->string('image')->nullable();
             $table->text('description');
-            $table->integer('category_id');
+            $table->bigInteger('category_id');
             $table->timestamps();
         });
     }
