+++ database/migrations/2022_04_28_143059_create_product_images_table.php
@@ -16,7 +16,7 @@ class CreateProductImagesTable extends Migration
         Schema::create('product_images', function (Blueprint $table) {
             $table->id();
             $table->string('image_path')->nullable();
-            $table->integer('product_id');
+            $table->bigInteger('product_id');
             $table->timestamps();
         });
     }
